backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
firebase-admin==6.4.0
pydantic==2.6.3
python-dotenv==1.0.1
httpx==0.27.2
openai==1.12.0
langgraph[visualization]==0.0.30
langchain==0.1.9
langchain-openai==0.0.8
ipython==7.34.0
grandalf==0.7
tavily-python==0.3.1